insert_tasks_ignore_duplicate_names - Moved insert function to new file in extra package.
